pyproject.toml
[tool.poetry]
name = "p9"
version = "0.1.0"
description = "CONCEVOIR ET DEPLOYER SYSTEME RAG"
authors = ["RomaneFatima-Zahra <barhoufz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
numpy = "^2.3.4"
pandas = "^2.3.3"
mistral = "^21.0.0"
mistralai = "^1.9.11"
ragas = "^0.3.9"
faiss-cpu = "^1.12.0"
langchain = "^1.0.5"
python-dotenv = "^1.2.1"
sentence-transformers = "^5.1.2"
langchain-mistralai = "^1.0.1"
fastapi = "^0.121.3"
uvicorn = "^0.38.0"
pytest = "^9.0.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^24.8.0"
flake8 = "^7.1.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
